# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TokHub External API
  version: 0.1.0
  summary: Public monitoring, status Open API, and dedicated gateway contracts.
  description: |
    This contract covers stable external integration points only.
    Browser session APIs under /api/auth, /api/me, /api/console and /api/admin are documented in docs/API.md.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
  - url: https://tokhub.example.com
    description: Production example
...
security: []
paths:
...
  /api/public/overview:
    get:
      tags: [Public Monitoring]
      summary: Public monitoring overview
      operationId: getPublicOverview
      responses:
        "200":
          description: Aggregated channel health and cost metrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Overview"
        "429":
          $ref: "#/components/responses/Error"
  /api/public/channels:
    get:
      tags: [Public Monitoring]
      summary: Public channel list
      operationId: listPublicChannels
      parameters:
        - $ref: "#/components/parameters/Dimension"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Provider"
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paged channel list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelList"
        "429":
          $ref: "#/components/responses/Error"
  /api/public/channels/{channelID}:
    get:
      tags: [Public Monitoring]
      summary: Public channel detail
      operationId: getPublicChannel
      parameters:
        - $ref: "#/components/parameters/ChannelID"
      responses:
        "200":
          description: Channel detail, current status and latest probe summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/public/channels/{channelID}/series:
    get:
      tags: [Public Monitoring]
      summary: Channel time series
      operationId: getPublicChannelSeries
      parameters:
        - $ref: "#/components/parameters/ChannelID"
        - name: days
          in: query
          schema:
            type: integer
            enum: [1, 7, 30, 90]
            default: 7
      responses:
        "200":
          description: Time series points.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SeriesPoint"
...
  /api/public/recommend/click:
    post:
      tags: [Public Monitoring]
      summary: Track recommendation click
      operationId: trackRecommendClick
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemType, itemId]
              properties:
                itemType:
                  type: string
                  enum: [pick, rank, reward, scenario, cta]
                itemId:
                  type: string
                channelId:
                  type: string
      responses:
        "202":
          description: Click accepted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: tracked
        "400":
          $ref: "#/components/responses/Error"
  /v1/status/overview:
    get:
      tags: [Status Open API]
      summary: Authenticated status overview
      operationId: getStatusOverview
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      responses:
        "200":
          description: Same overview data as the public dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Overview"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/status/channels:
    get:
      tags: [Status Open API]
      summary: Authenticated status channel list
      operationId: listStatusChannels
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      responses:
        "200":
          description: Channel list scoped by Site Key permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelList"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/status/channels/{channelID}:
    get:
      tags: [Status Open API]
      summary: Authenticated status channel detail
      operationId: getStatusChannel
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/ChannelID"
      responses:
        "200":
          description: Channel detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/status/channel-sync:
    get:
      tags: [Status Open API]
      summary: TokHub channel sync package
      description: Requires a Site Key with the channel_sync scope. Credentials are only included when includeCredentials=1 is requested.
      operationId: getStatusChannelSync
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      parameters:
        - name: includeCredentials
          in: query
          schema:
            type: string
            enum: ["0", "1"]
            default: "0"
          description: Set to "1" only for trusted TokHub-to-TokHub synchronization; response then contains provider API keys.
      responses:
        "200":
          description: Platform channel configuration, monitor snapshot, and recent probe logs for synchronization.
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always no-store.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSyncPackage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/status/uptime:
    get:
      tags: [Status Open API]
      summary: Authenticated uptime summary
      operationId: getStatusUptime
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      responses:
        "200":
          description: Uptime summary and trends.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UptimeSummary"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/status/incidents:
    get:
      tags: [Status Open API]
      summary: Authenticated incident list
      operationId: listStatusIncidents
      security:
        - SiteKeyHeader: []
        - SiteKeyBearer: []
      responses:
        "200":
          description: Incident list.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /gateway/v1/models:
    get:
      tags: [Gateway]
      summary: OpenAI-compatible model list
      operationId: listGatewayModels
      security:
        - GatewayBearer: []
      responses:
        "200":
          description: OpenAI-compatible models response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayModels"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /gateway/v1/chat/completions:
    post:
      tags: [Gateway]
      summary: OpenAI-compatible chat completions
      operationId: createGatewayChatCompletion
      security:
        - GatewayBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: OpenAI-compatible completion or stream.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /gateway/v1/responses:
    post:
      tags: [Gateway]
      summary: OpenAI Responses-compatible endpoint
      operationId: createGatewayResponse
      security:
        - GatewayBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseRequest"
      responses:
        "200":
          description: OpenAI-compatible Responses payload or stream.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    SiteKeyHeader:
      type: apiKey
      in: header
      name: X-Site-Key
    SiteKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: site_key
    GatewayBearer:
      type: http
      scheme: bearer
      bearerFormat: gateway_key
  parameters:
    ChannelID:
      name: channelID
      in: path
      required: true
      schema:
        type: string
      example: ch_openai_gpt4o
    Dimension:
      name: dimension
      in: query
      schema:
        type: string
        enum: [brand, model]
        default: brand
    Status:
      name: status
      in: query
      schema:
        type: string
        enum: [healthy, degraded, down, auth_error, unknown]
    Provider:
      name: provider
      in: query
      schema:
        type: string
    Query:
      name: query
      in: query
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    Error:
      description: Controlled TokHub error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Overview:
      type: object
      additionalProperties: true
      properties:
        total:
          type: integer
        healthy:
          type: integer
        degraded:
          type: integer
        down:
          type: integer
        authError:
          type: integer
        p95LatencyMs:
          type: integer
        dailyCostUsd:
          type: number
    ChannelList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    ChannelDetail:
      allOf:
        - $ref: "#/components/schemas/Channel"
        - type: object
          additionalProperties: true
          properties:
            l1:
              type: array
              items:
                $ref: "#/components/schemas/ProbeStep"
            l2:
              type: array
              items:
                $ref: "#/components/schemas/ProbeStep"
            l3:
              type: object
              additionalProperties: true
...